move everything to one place
