+++ src/assets/ka.png
